fix(capture): insert the backfill run and its job atomically
