+++ backend/internal/compose/backfilltransport.go
@@ -18,6 +18,7 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	openapi_types "github.com/oapi-codegen/runtime/types"
 	"github.com/riverqueue/river"
@@ -251,29 +252,39 @@ func (h backfillHandlers) StartConnectorBackfill(w http.ResponseWriter, r *http.
 	if messages, err := h.registry.EstimateBackfill(r.Context(), string(provider), userID, months); err == nil {
 		estimate = messages
 	}
-	run, err := h.registry.StartBackfill(r.Context(), string(provider), userID, months, estimate)
-	if err != nil {
-		h.writeBackfillError(w, r, err)
-		return
-	}
 	ws, ok := principal.WorkspaceID(r.Context())
 	if !ok {
-		// StartBackfill just committed under a workspace-bound transaction, so
-		// a missing workspace here is a wiring defect, surfaced honestly.
+		// Every authenticated request carries its workspace, so its absence is a
+		// wiring defect — surfaced before anything is written, since the job the
+		// run needs cannot be addressed without it.
 		httperr.Write(w, r, &httperr.DetailedError{
 			Status: http.StatusInternalServerError, Code: "workspace_missing",
-			Detail: "The request carries no workspace context; the run was recorded but not scheduled. Try again.",
+			Detail: "The request carries no workspace context; nothing was started. Try again.",
 		})
 		return
 	}
-	if err := h.inserter.Enqueue(r.Context(), CaptureBackfillArgs{
-		Workspace: ws.String(), BackfillID: run.ID.String(),
-	}, &river.InsertOpts{UniqueOpts: river.UniqueOpts{ByArgs: true, ByState: activeSweepStates}}); err != nil {
-		h.log.ErrorContext(r.Context(), "backfill: enqueue", "err", err)
+	// The job is inserted inside the run's transaction, so an unreachable queue
+	// leaves no queued row for the live-run index to make permanent. enqueueErr
+	// is kept out of the closure's return path so the failure keeps its own
+	// status and copy instead of the generic connector-fault mapping.
+	var enqueueErr error
+	run, err := h.registry.StartBackfill(r.Context(), string(provider), userID, months, estimate,
+		func(ctx context.Context, tx pgx.Tx, backfillID ids.UUID) error {
+			enqueueErr = h.inserter.EnqueueTx(ctx, tx, CaptureBackfillArgs{
+				Workspace: ws.String(), BackfillID: backfillID.String(),
+			}, &river.InsertOpts{UniqueOpts: river.UniqueOpts{ByArgs: true, ByState: activeSweepStates}})
+			return enqueueErr
+		})
+	if enqueueErr != nil {
+		h.log.ErrorContext(r.Context(), "backfill: enqueue", "err", enqueueErr)
 		httperr.Write(w, r, &httperr.DetailedError{
 			Status: http.StatusInternalServerError, Code: "backfill_enqueue_failed",
-			Detail: "The backfill was recorded but could not be scheduled. Try again — the run resumes, nothing is lost.",
+			Detail: "The backfill could not be scheduled, so nothing was started. Try again.",
 		})
+		return
+	}
+	if err != nil {
+		h.writeBackfillError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusAccepted)
